+++ app/Http/Controllers/RoutineWomenVisitController.php
@@ -35,7 +35,7 @@ class RoutineWomenVisitController extends Controller
 
         $validator = Validator::make($request->all(), [
             'medical_record_id' => ['required', 'integer', 'exists:medical_records,id'],
-            'z_score' => ['required', 'integer'],
+            'z_score' => ['required', 'numeric'],
             'current_status' => [
                 'required',
                 Rule::in(['mam','normal']),
